refactor: extract app_builders from api/app.py and fix test imports
